Renamed files to be consistent and fixed imports
