+++ commandchecks.py
